requirements.txt
asgiref==3.5.2
attrs==22.1.0
Django==4.0.6
django-safedelete==1.2.2
djangorestframework==3.13.1
drf-spectacular==0.23.1
inflection==0.5.1
jsonschema==4.9.0
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
sqlparse==0.4.2
uritemplate==4.1.1
